Disambiguate items in batch processing for sessions
The creation of sessions allows to submit a set of them to treat them as
an aggregate when requesting to operate on a certain peering posture.
This commit adds a field to the input and output series such that the
client can univocally identify which sessions have been processed
successfully or not without relying on implicit ordering of the input or
result set.