Add deletion of individual sessions
Implement a method to depeer by deleting existing sessions. The
batched creation method does not create an obvious collection resource
on the server side that may be used for a depeering situation therefore
this commit only implements individual session deletion.

To tackle a situation where depeering is desired for a whole metro or
IX while respecting this RESTful interface, we will need to define a
separate aggregate resource that can be used to operate atomically on
the set of target sessions.