Remove transient identifiers and partial success in response
As per the different discussions, this commit removes references to
identifiers used to identify parts of an inflight request that
disambiguate failure scenarios. They may be addressed later.

It also removes the partial successful reponse conveyed through a
redirect error code. The rationale to use only the 200 success code
is that the response for a batched request has been successfully
processed and the response may still contain references to items which
were not successful. Clients will have to introspect the response to
decide what to do with those failed items.